Update Data.php to fix BasePrice on config Product
Magneto deliver different prices for $product->getFinalPrice() or $product->getPriceInfo()->getPrice('final_price')->getValue();
The Problem is for config products on getFinalPrice the "Dummy" price of parent is delivert and not the minimum price of the children.
"Dummy" price means when you create an config product you have to set an price for the parent because it is an requiered field. Normaly the "Dummy" price  is not shown on frontend.